Use in console.timeStamp (#54339)
Summary:

# Changelog: [Internal]

Whenever the `console.timeStamp()` is called and we are actually tracing, we will peek at the `ConsoleTaskOrchestrator` stack and check if there is a task.

If so, it means that this call happened inside a task context, so we can request a stack trace and propagate it down to `PerformanceTracer`.

Differential Revision: D85481866