Fix keyboard navigation through nested a11y views in ReactHorizontalScrollView with snapToInterval enabled (#54351)
Summary:

Navigating with keyboard through a11y items in horizontal FlatList that are not the direct child of the ScrollView contentView is broken, because instead of focusing the next focusable sibling, arrowScroll scrolls to the next page. This issue only happens when FlatList has snapToInterval enabled.
Fixed the condition that checked if a given focusable view belongs to the contentView.

Changelog: [Internal]

Reviewed By: Abbondanzo

Differential Revision: D85261167