Fix missing enum values in switch statements (#54373)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/54373

`-Wswitch-enum` was introduced in 2024 and is beneficial because it will err when switch statement is missing a case for an enum, even with `default:` present.  This helps alert developers when they add an enum value of all the switch statements that need updating.

These diffs are to help progress the codebase so that we can enable `-Wswitch-enum` by default in `fbobjc`

## Changelog:
[Internal] - Make enum exhaustive

Reviewed By: cipolleschi

Differential Revision: D85956869

fbshipit-source-id: 2833822e1f82e77909229c566798547e1cb7f084