Fix keyboard navigation through nested a11y views in ReactHorizontalScrollView with snapToInterval enabled (#54351)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/54351

Navigating with keyboard through a11y items in horizontal FlatList that are not the direct child of the ScrollView contentView is broken, because instead of focusing the next focusable sibling, arrowScroll scrolls to the next page. This issue only happens when FlatList has snapToInterval enabled.
Fixed the condition that checked if a given focusable view belongs to the contentView.

Changelog: [Android] [Fixed] - Fix keyboard navigation through items in horizontal ScrollView with snapToInterval enabled

Reviewed By: Abbondanzo

Differential Revision: D85261167

fbshipit-source-id: 220c6648c0381899f312b1b1f3923ed5203899b1