Maintain momentum through non-zero velocity when paging is enabled (#54413)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/54413

We should treat `ACTION_SCROLL` events akin to momentum scrolling--they represent a scroll with non-zero velocity--but not as momentum events themselves--they don't have a real velocity after they've dispatched. When paging is enabled, we should run the same `flingAndSnap` logic and rely on this non-zero velocity to encourage scrolling to the next item in the direction of the scroll rather than immediately picking the nearest offset.

The end effect of this change is that scrolling down will always go to the next item regardless of when scrolling stops. Scrolling up will always go to the previous item. Similar for horizontal scrolling per right/left scrolling.

Changelog: [Internal]

Reviewed By: necolas

Differential Revision: D86151012

fbshipit-source-id: 14987beac931e3d47a4cc013f5d2a4ccb5c2b8bd