No-op by default (#54335)
Summary:

# Changelog: [Internal]

Actually gate the task scheduling logic, so that the implementation is a no-op by default, if we are not tracing in the background or have Runtime CDP domain enabled.

The drawback is that we might loose stack traces for some component renders, if the corresponding JSX declaration was called before we've enabled the `console.createTask()`.

Reviewed By: huntie

Differential Revision: D85849862