Serialize stack trace eagerly
Summary:
# Changelog: [Internal]

Previosuly, we were serializing only at the end of tracing. We were creating a weak pointer to encapsulating `ConsoleTaskContext`.

Initially I've decided to do it this way for performance reasons, but this resulted in missing a significant portion of stack traces, if the corresponding task was already gc-ed.

Unfortunately, we have to serialize them at the time of reporting to `PerformanceTracer`.

Reviewed By: huntie

Differential Revision: D86207507