Back out "Remove RCT_EXPORT_MODULE from Core modules"
Summary:
Backing this out because there are a couple of internal tests going in timeout when this land.

I believe that for the tests some of these modules are not registered correctly.

## Changelog:
[Internal] -

## Facebook:

See T244088155. For some reasons, these two tests:

- `buck test @//fbobjc/mode/buck2/ios-tests fbsource//fbobjc/Libraries/FBReactKit:FDSMediaWrapperServerSnapshotTests_Facebook` ([link](https://www.internalfb.com/intern/test/844425131558081))
- `buck test @//fbobjc/mode/buck2/ios-tests fbsource//fbobjc/Libraries/FBReactKit:FDSInlineReactionsServerSnapshotTests_Facebook` ([link](https://www.internalfb.com/intern/test/844425131558078))

timeouts with diff D85580258. I need to investigate which modules make these tests time out. It could be that the test setup is missing some plugin.

I'm also surprised why other FDS tests are actually passing...

Differential Revision: D86401358