Serialize stack trace eagerly (#54418)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/54418

# Changelog: [Internal]

Previosuly, we were serializing only at the end of tracing. We were creating a weak pointer to encapsulating `ConsoleTaskContext`.

Initially I've decided to do it this way for performance reasons, but this resulted in missing a significant portion of stack traces, if the corresponding task was already gc-ed.

Unfortunately, we have to serialize them at the time of reporting to `PerformanceTracer`.

Reviewed By: huntie

Differential Revision: D86207507

fbshipit-source-id: 58b15145ef081a02f2d79d75e05380968eb43a06