Define main entities (#54334)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/54334

# Changelog: [Internal]

This defines the main entities for the console.createTask():
- ConsoleTaskOrchestrator: global stack that can be used as a source of pending tasks.
- ConsoleTaskContext: RAII object, captures the context for a specific task. Lifetime is bound to the lifetime of the task object in JavaScript.
- ConsoleTask: RAII-like object. Initialized only during the callback run of `task.run(...)`.

I couldn't find a better way to solve this without having a static singleton. Native modules don't have access to ReactInstance object, so we won't be able to access this global stack from `performance.measure` implementation, for example.

Not using the word `async` anywhere in the naming, because the current implementation doesn't support async stack traces.

Reviewed By: sbuggay

Differential Revision: D85481864

fbshipit-source-id: 4a0b317cccb8c2492647f06472b00fc0daa2c8a3