Use in console.timeStamp (#54339)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/54339

# Changelog: [Internal]

Whenever the `console.timeStamp()` is called and we are actually tracing, we will peek at the `ConsoleTaskOrchestrator` stack and check if there is a task.

If so, it means that this call happened inside a task context, so we can request a stack trace and propagate it down to `PerformanceTracer`.

Reviewed By: huntie

Differential Revision: D85481866

fbshipit-source-id: 5bb7c4b7505f0811eff91eed53dc7822a0c063eb