Fix cascading update (#54438)
Summary:

# Changelog: [Internal]

`LogBoxStateSubscription` subscribed to state updates in a mount effect. This callback could be called instantly on subscription initiation, which would cause a cascading update => setState in mount effect [1].

This diff changes the approach, the state is initialized before mount and the callback is not called synchronously on initialization. I had to keep `observe` function as it is right now, because it is used extensively in tests.

Reviewed By: javache

Differential Revision: D86449266