Added Fantom testing framework rules file for future devmate sessions
Summary: Adding a Fantom.md rule file as an example for the org to work with evals. This rules file just has some basics, largely distilled from the Fantom docs by Devmate. It's not guaranteed to be super great yet, we'll use evals further up in the stack to start to improve these rules based on pre-existing tests.

Differential Revision: D85880548