Add runtime validation of Props 2.0 output (#54461)
Summary:

Compare the output from Props 1.5 with Props 2.0 to detect any missing props in the output coming from Props 2.0.

The comparison doesn't run for initial mounts due to Props 1.5 not using the default property values to generate a diff, which is obviously leading to more properties being in the Props 1.5 output that aren't necessary.

Changelog: [Internal]

Differential Revision: D86577390