Fix jagged edge rendering on API 28 (#54525)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/54525

Older versions of Android's clipPath APIs do not properly support antialiasing and BackgroundStyleApplicator relies on this API to clip to border radii that are smaller than provided images or views. As a result, this often leaves jagged edges on the outer edges of these images. Android even calls this out in their documentation under the "Canvas.clipPath" section of https://developer.android.com/topic/performance/vitals/render#common-jank.

This change introduces a new API in BackgroundStyleApplicator called `clipToPaddingBoxWithAntiAliasing` which takes an additional argument over the existing `clipToPaddingBox`: a draw call. If non-null and the Android API is 28 or lower, a new drawPath operation is used to mark the pixels outside of the drawn contents as transparent.

Changelog: [Android][Fixed] - Fixed antialiasing issues from border radius usage in API <= 28

Reviewed By: haixia-meta, lenaic

Differential Revision: D86263195

fbshipit-source-id: 5dba744da08a3c9297f921baf7d4f1a354404f74