Adding support for unions in Codegen Schema (#54489)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/54489

Previously, the parser was mapping to Array<any> when the element types of union were different. This caused compat checker to not recognize changes in unions of different types.

`StringLiteralUnionTypeAnnotation`  already had support for `types`, that can be expanded to all Union types so as to incorporate the Unions of different types.
```
{
  type: 'StringLiteralUnionTypeAnnotation',
  types: [
    {
      type: 'StringLiteralTypeAnnotation'
      value: 'light'
    },
    {
      type: 'StringLiteralTypeAnnotation'
      value: 'dark'
    },
  ],
}
```

NOTE: Generators behavior still remain unchanged hence don't allow unions of different types. Only RN Compat Checker consumes this.

The value of this is that the compat checker can now error when the union of different types changes.
Also for consistency added support for similar literal types for Number & Boolean i.e `NumberLiteralTypeAnnotation` & `BooleanLiteralTypeAnnotation`

Changelog: [Internal]

Differential Revision: D86501597