Fix create-draft-release syntax error
Summary:
This change fix a syntax error in the create-draft-release workflow. We were basically passing the Hermes and HermesV1 versions without wrapping them in `'`.

JS was interpreting them as numbers and making a mess of the syntax.

Wrapping them in `'` should fix the job.

The commit was already picked in the release branch.

## Changelog:
[Internal] -

Differential Revision: D87244325