Fix create-draft-release syntax error (#54567)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/54567

This change fix a syntax error in the create-draft-release workflow. We were basically passing the Hermes and HermesV1 versions without wrapping them in `'`.

JS was interpreting them as numbers and making a mess of the syntax.

Wrapping them in `'` should fix the job.

The commit was already picked in the release branch.

## Changelog:
[Internal] -

Reviewed By: cortinico

Differential Revision: D87244325

fbshipit-source-id: b115c233779b36a1b6b0462eb0a8ef60ceafce7c