Add support for UIAccessibilityPriority (#54289)
Summary:
Adds support for UIAccessibilityPriority on iOS. This is important because we want to be able to set high-priority announcements that cannot be interrupted.

Use case: a modal opens, user performs an action that closes the modal and returns to the previous page. We need to announce it but the voiceover engine will target an element as we've just navigated and cancel the announcement because the default announcement priority is interruptible.

https://developer.apple.com/videos/play/wwdc2023/10036/?time=293 Here's a talk about this in wwdc 2023.
<img width="1813" height="981" alt="Screenshot 2025-10-27 at 11 12 18" src="https://github.com/user-attachments/assets/8306d6f6-a5a9-4834-a5d9-c5b2e014c258" />

## Changelog:

[IOS][ADDED] - Add support for UIAccessibilityPriority in announcements

Pull Request resolved: https://github.com/facebook/react-native/pull/54289

Test Plan:
Open the rn-tester app, go to Accessibility -> Check if the screen reader announces.

Press: Announce for accessibility High Priority. You will hear "High Priority Announcement Test, Normal Priority Announcement". The second announcement did not cancel the first announcement (expected behavior)

Press: Announce for accessibility Low Priority: You will only hear "Normal Priority Announcement". The low priority announcement was successfully interrupted.

Reviewed By: cipolleschi

Differential Revision: D87477523

Pulled By: javache

fbshipit-source-id: 5e8ccf26188cce1407cb814df6993fcbf1ade7ce