Use new API for tracing state on Android (#54629)
Summary:

# Changelog: [Internal]

Before this change, every `getTracingState()` call would go through jni to C++ to get the current state synchronously.

This has a risk of data races, since we can't guarantee from which thread the Host will call this method. At the same time, the Tracing could be started from inspector thread, if this was initiated with CDP request.

Instead, the state will be propagated asynchronously via `HostTargetTracingDelegate`, and then it will be encapsulated on the Host side. In android case, this will be part of the `ReactHostInspectorTarget`.

Also removing no longer used code in `jsinspector-modern`.

Reviewed By: sbuggay

Differential Revision: D87123722