Introduce TimeWindowedBuffer
Summary:
# Changelog: [Internal]

Adding a data structure that could be configured to preserve entries that are inside of a specified time window.

At the same time, this data structure could be used as a simple buffer, if no arguments were specified during constructing.

This follows the approach that was added to `PerformanceTracer`, but it doesn't use pointers to avoid dealing with updates during copies or moves.

Differential Revision: D87776299