Keep tracing time window on TraceRecordingState
Summary:
# Changelog: [Internal]

`HostTargetTraceRecording` will receive the tracing window size as a parameter and will propagate it to `TraceRecordingState`.

This will make sure that all Tracing Agents actually aware of the window size.

For example, once we add Frames to Host layer, it should use the same window size, as `PerformanceTracer`, which is initialized in `InstanceTracingAgent`.

Differential Revision: D87782664