Apply categories to changelog
Summary:
This change applies categories to the changelog.

## Changelog:
[Internal] -

## Facebook:

I used AI to do so, and it did a pretty good job.

Prompt:
> if you look at the CHANGELOG.md for versions like v0.82.0, v0.81.0, v0.80.0, we have been annotating all the commits with an initial tag, which represent the area affected by the commit.
> I want you to go through all the commits of v0.83.0-rc.3, v0.83.0-rc.2, v0.83.0-rc.1, and v0.83.0-rc.0 and annothate each commit with the corresponding area.
> Try to reuse areas from previous releases if you can find matches. If it is unclear which area is affected, you can find the link to the commit at the end of the entry. Try to navigate there and read the commit summary to understand which area is affecting.

This is [the trajectory](https://www.internalfb.com/intern/devai/devmate/investigator?paste=P2056034816)

One **limitation** of DevMate is that it can't go and see what the commit is actually about. Probably cursor can do that.

Reviewed By: cortinico

Differential Revision: D87870198