Fix RCTDevLoadingView crash when adding button constraints (#54690)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/54690

Changelog:
[iOS][Fixed] - Fixed crash from dismiss button in DevLoadingView

D86420230 added a dismiss button feature to RCTDevLoadingView. However, when adding layout constraints on line 198, it incorrectly checked the `dismissButton` parameter instead of the `self->_dismissButton` instance variable. This caused NSLayoutConstraint crashes when the parameter was nil/false but the instance variable existed from a previous call, or vice versa.

This changes line 198 to check `self->_dismissButton` (the instance variable) instead of `dismissButton` (the function parameter) to properly verify if the button exists before adding constraints to it.

Reviewed By: vzaidman, javache

Differential Revision: D87871856

fbshipit-source-id: c8fe033bce383c2589041c9cd3fad65f37553d3c