Define FrameTimingSequence (#54674)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/54674

# Changelog: [Internal]

This diff introduces a definition for `FrameTimingSequence`, which will represent the different phases of a frame that will be reported as part of the trace.

These sequences will be stored on a `HostTargetTraceRecording`. It is `Host`'s responsibility to propagate these entities to `HostTarget`, there will be a corresponding method for this.

Reviewed By: sbuggay

Differential Revision: D87373811

fbshipit-source-id: bd894f8c39a6d4ee0bef1303e4f801b27fabc901