Introduce TimeWindowedBuffer (#54679)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/54679

# Changelog: [Internal]

Adding a data structure that could be configured to preserve entries that are inside of a specified time window.

At the same time, this data structure could be used as a simple buffer, if no arguments were specified during constructing.

This follows the approach that was added to `PerformanceTracer`, but it doesn't use pointers to avoid dealing with updates during copies or moves.

Reviewed By: sbuggay

Differential Revision: D87776299

fbshipit-source-id: d20bf6dd08af7971aaf34ffc11ec5993b8edaf4b