Keep tracing time window on TraceRecordingState (#54673)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/54673

# Changelog: [Internal]

`HostTargetTraceRecording` will receive the tracing window size as a parameter and will propagate it to `TraceRecordingState`.

This will make sure that all Tracing Agents actually aware of the window size.

For example, once we add Frames to Host layer, it should use the same window size, as `PerformanceTracer`, which is initialized in `InstanceTracingAgent`.

Reviewed By: sbuggay

Differential Revision: D87782664

fbshipit-source-id: 58203065d68ac898adc785317ec2cb792efd13dc