Clarify HostTracingProfile and HostTraceRecordingState (#54677)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/54677

# Changelog: [Internal]

Currently, `HostTraceRecordingState` is used everywhere and it is the final data struct that contains the tracing profile.

At the same time, the original idea for it was to be a global state for Tracing Agents, where they could:
1. Read the tracing configuration, like tracing mode, enabled categories, etc.
2. Stash corresponding target profiles on reloads. For example, if `RuntimeTarget` was reloaded during tracing, `RuntimeTracingAgent` would stop JavaScript sampling profiler, and then stash this profile on a state.

This change should clarify the data structures and when they are used.

Reviewed By: sbuggay

Differential Revision: D87830455

fbshipit-source-id: 15c48baede85a2d85744d9c3c3b7a419c1f1d299