fix(a11y): android prevent redundant double screen reader focus step on `<Text role="link">` (#54515)
Summary:
Resolves https://github.com/facebook/react-native/issues/53240

Resolves an Android-only issue where a top-level `<Text role="link">Link Text</Text>` would generate two accessibility (screen reader) focus steps:
1. "Link Text. Link. Links available, use tap with 3 fingers to view."
2. "Link Text. Link. Double-tap to activate."

This behavior was inconsistent with iOS behavior, which generated a single screen reader step, resulting in a poor user experience due to unnecessary repetition of information.

This PR resolves the issue in a way that allows mixing plain text and links while still yielding additional focus steps for nested links if there is any other text present or if there are multiple links.

## Android after

https://github.com/user-attachments/assets/646b64af-a7eb-4b28-8735-546b9909a510

## Android before

https://github.com/user-attachments/assets/43bd9ee9-890f-43ca-9363-ade8b45dea27

## iOS (for comparison)

https://github.com/user-attachments/assets/807d1771-e3e7-4dca-9409-7697d1a0cb9c

## Changelog:

<!-- Help reviewers and the release process by writing your own changelog entry.

[ANDROID] [FIXED] a11y: prevent redundant double screen reader focus steps on `<Text role="link">`
Pick one each for the category and type tags:

[ANDROID|GENERAL|IOS|INTERNAL] [BREAKING|ADDED|CHANGED|DEPRECATED|REMOVED|FIXED|SECURITY] - Message

For more details, see:
https://reactnative.dev/contributing/changelogs-in-pull-requests

Pull Request resolved: https://github.com/facebook/react-native/pull/54515

Test Plan: Open RN Tester on "Text with link role" example and test with screen reader.

Reviewed By: cipolleschi

Differential Revision: D87464727

Pulled By: joevilches

fbshipit-source-id: a69df5b63128886ea5c45b22e068d455ee808c0f